// apps/api/src/common/enums/user-action.enum.ts
export enum UserAction {
    USER_ACTION = "USER_ACTION",

    CREATE = "CREATE",

    UPDATE = "UPDATE",

    DELETE = "DELETE",

    RESTORE = "RESTORE",

    ACTIVATE = "ACTIVATE",

    DEACTIVATE = "DEACTIVATE",

    BAN = "BAN",

    UNBAN = "UNBAN",

    CHANGE_ROLE = "CHANGE_ROLE",

    VERIFY_EMAIL = "VERIFY_EMAIL",

    RESET_PASSWORD = "RESET_PASSWORD",
}